up code

+++ modules/admin/products/views/indexView.php
@@ -57,7 +57,7 @@
                      <td>
                         <a href="#"><button> <i class="fa-solid fa-pen mr-2"></i></button></a>
 
-                        <a href="?role=admin&mod=products&action=delete&id=<?php echo $row['id'] ?>">
+                        <a href="?role=admin&mod=products&action=delete&id=<?php echo $row['id'] ?>"onclick="return confirm('Bạn chắc chắn muốn xoá danh mục: <?php echo $row['name'] ?> không? Hành động sẽ xoá danh mục và toàn bộ sản phẩm có trong danh mục này.')">
                            <button><i class="fa-solid fa-trash-can"></i></button>
                         </a>
                      </td>
